Decrease observer calls
This commit avoids to use three times of the same observer on three
properties and sets one observer for all of those three properties. This
reduces the number of calls. For example, on the demo page, the observer
was triggered 25 times instead of 9 times with the new version. As you
can see, the default value of the date is the current one.

Closes #15