feat(locale): add shared locales
This commit adds a property "locale" to the "moment-js" element. When
this property changes, the according language file is loaded (if it's
not already done) and the shared property "MomentJs.language" is
updated. All instance of the element refreshes its display when the
language file is imported. By default they use the navigator language
or English if this one is not found.

Closes #3