fix(demo): use webcomponentsjs loader
Before this commit the demo page was importing the webcomponentsjs
minified file. The lint step was failling because since v1 of
webcomponentsjs package there no minified version anymore. It was
replaced by a loader. Now, the demo page is using this loader.

The rules used by the linter was also changed from polymer-1 to hyrbid
element.