chore(i18n): improve mechanism of language sharing
This commit updates the mechanism of locales file loading by
replacing the simple function by a Promise. It ensures the locales are
loaded only once.

Moreover, now instead of asking to each moment-js instances to update
their 'lang' property, the property is updated by the element that has
loaded the locales file. Doing so, when the observer is run the element
knows if it has to load the locales or not (because it property lang is
alread set, the file is already loaded).

Finally, I also remove the warning about loading locales which does not
exist from the demo page because it's not an problem anymore. When the
file is not found, the moments do not change language. They'll continue
to follow the next locales changes.