fix(lifecycle): attach listener before MomentJs is ready
Before this commit, the listeners to initialize the element when
MomentJs is ready were attached only when the element itself was ready.
Sometimes MomentJs was ready before the element and the element was not
properly initialized.

Now, the listeners are attached when the element is created and before
it is ready and attempts to load MomentJs. So elements created on the
fly are well initialized. The listeners are now passive.

fixes #39